chore: split up tauri-conf-schema.ts to make it more readable and manageable
