+++ web-client/src/lib/tauri/build-schema-map.ts
@@ -1,0 +1,20 @@
+import { Draft07, JsonSchema, JsonPointer } from "json-schema-library";
+
+export function buildSchemaMap(baseSchema: JsonSchema, data: object) {
+  const jsonSchema = new Draft07(baseSchema);
+  const map = new Map();
+
+  const buildMap = (
+    schema: JsonSchema,
+    value: unknown,
+    pointer: JsonPointer
+  ) => {
+    //schema = jsonSchema.compileSchema(schema);
+    pointer = pointer.replace("#/", "").replace("#", "").replaceAll("/", ".");
+    map.set(pointer, schema);
+  };
+
+  jsonSchema.each(data, buildMap);
+
+  return map;
+}
